Changes made in Sept 9

+++ app/src/main/java/com/clxns/app/data/model/LoginResponse.kt
@@ -49,11 +49,11 @@ data class LoginData(
     @Json(name = "lastName")
     val lastName: String,
     @Json(name = "latitude")
-    val latitude: Int,
+    val latitude: Double,
     @Json(name = "location")
     val location: String,
     @Json(name = "longitude")
-    val longitude: Int,
+    val longitude: Double,
     @Json(name = "otp")
     val otp: Any?,
     @Json(name = "otpExpired")
